KAFKA-14683 Migrate #testPartitionCountInCaseOfPartitionRevocation to Mockito
